1. c化bug修订，修复bug如下 1.1. api_event_db_dumper.cpp: 修复落盘task和api无connectionId 1.2. chunk_generator.cpp：修复文件读取排序错误，修改前按字符串排序会导致文件乱序 1.3. event.h: id作为connectionId来源支持负值 1.4. event_grouper.cpp：全量API都生成event，用于生成connectionId 1.5. file.cpp: 新增文件排序函数及basename获取函数 1.6. host_trace_worker.cpp: 全量event排序，避免老化非老化数据混合场景 1.7. time_logger.h: 避免不必要的打屏 1.8. tree_analyzer.cpp: 增加connectionId落盘逻辑；修复ctxId拷贝错误bug；增加计算类算子ts_track数据任务类型白名单； 1.9. tree_analyzer_utest.cpp: 修复错误的场景1信息，新增场景2 L0及L1级别用例
2.功能增强：
2.1. event_queue.cpp: 排序时间统计
2.2. tree_builder.cpp: 维测信息增强
2.3. tree_builder_utest.cpp: 增加更灵活的event插入方式


# message auto-generated for no-merge-commit merge:
merge bugfixc into master

c化bug修订及功能增强

Created-by: msprofDeveloper

Match-id-25c94f2da9b572ea91717403ebf2f14cd5f5a41c

+++ analysis/csrc/entities/event.h
@@ -81,7 +81,7 @@ struct Event {
         std::shared_ptr<ConcatTensorInfo> tensorPtr;
     };
     EventInfo info;
-    uint32_t id = 0;    // 全局唯一ID
+    int64_t id = 0;    // 全局唯一ID
 
     ~Event()
     {
